+++ WPayMaster/ViewModel/AdditionalViewModel/EditUserViewModel.cs
@@ -3,7 +3,6 @@ using System.Windows;
 using System.Windows.Input;
 using DataBaseService;
 using DataBaseService.Context;
-using DataBaseService.Model;
 using Microsoft.Win32;
 using PropertyChanged;
 using Shared;
@@ -29,7 +28,7 @@ namespace ViewModel.AdditionalViewModel
         public string Sex { get; set; }
         public string Post { get; set; }
         public string Password { get; set; }
-        public int TariffRate { get; set; }
+        public double TariffRate { get; set; }
         public string ImagePath { get; set; }
 
         public List<string> UserPostList { get; set; }
@@ -74,7 +73,7 @@ namespace ViewModel.AdditionalViewModel
             Sex = null;
             Post = null;
             Password = " ";
-            TariffRate = 1;
+            TariffRate = 1.0;
         }
 
         private void Close()
